change some coordinate to integer

+++ SuperMarioBros/FirePlant.h
@@ -41,7 +41,7 @@ protected:
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
 	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	void Render();
-	bool MarioDetection(int x, float y);
+	bool MarioDetection(int x, int y);
 	//virtual int IsCollidable() { return 1; };
 	virtual int IsBlocking() { return 0; }
 	virtual void OnNoCollision(DWORD dt);
